+++ src/main/java/com/lowdragmc/photon/client/PhotonClientProxy.java
@@ -1,8 +1,6 @@
 package com.lowdragmc.photon.client;
 
 import com.lowdragmc.photon.PhotonCommonProxy;
-import com.mojang.datafixers.util.Pair;
-import net.minecraft.client.renderer.ShaderInstance;
 import net.neoforged.api.distmarker.Dist;
 import net.neoforged.api.distmarker.OnlyIn;
 import net.neoforged.bus.api.IEventBus;
@@ -10,7 +8,6 @@ import net.neoforged.bus.api.SubscribeEvent;
 import net.neoforged.fml.event.lifecycle.FMLClientSetupEvent;
 import net.neoforged.neoforge.client.event.RegisterShadersEvent;
 
-import java.util.function.Consumer;
 
 @OnlyIn(Dist.CLIENT)
 public class PhotonClientProxy extends PhotonCommonProxy {